lib/oelite: teach oe show to suppress dynvar replacement
When looking at "oe show" output, it can sometimes be a little
confusing that some variables are apparently not fully expanded, since
we do the dynvar back-replacement dance in MetaData::dump_var. It's
easy enough to let the user choose whether he wants that to happen:
just don't build the list of dynvars unless the new dynvar_replacement
arg to MetaData::dump is True (which it is by default).

The option is deliberately hidden from "oe show --help", since I
couldn't find a reasonable way to explain what the dynvar replacement
is about in a single line that would be understandable, so for now
it's reserved for those who read the source and/or commit messages.